requirements.txt
boto3==1.24.41
celery==5.2.7
coverage==6.4.2
dj-rest-auth==2.2.5
Django==3.2.14
django-admin-interface==0.19.1
django-allauth==0.51.0
django-celery-beat==2.3.0
django-cors-headers==3.13.0
django-environ==0.9.0
django-extensions==3.2.0
django-filter==22.1
django-object-actions==4.0.0
django-js-reverse==0.9.1
django-nested-inline==0.4.5
django-polymorphic==3.1.0
django-rest-polymorphic==0.1.10
django-silk==5.0.1
django-storages==1.12.3
django-structlog==2.2.0
djangorestframework==3.13.1
drf-spectacular==0.22.1
Faker==13.15.1
gevent==21.12.0
gunicorn==20.1.0
hashids==1.3.1
openpyxl==3.0.10
psycopg2-binary==2.9.3
pytest==7.1.2
pytest-cov==3.0.0
pytest-django==4.5.2
pytest-factoryboy==2.5.0
requests==2.28.1
sentry-sdk==1.8.0
structlog==22.1.0
Werkzeug==2.2.0
whitenoise==6.2.0
